requirements.txt
streamlit==1.32.2
streamlit-drawable-canvas==0.9.3
opencv-python==4.11.0.86
numpy==1.26.4
pillow==10.2.0
torch==2.7.0
torchvision==0.22.0
requests==2.32.3